requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.24.0
streamlit>=1.28.0
PyQt5>=5.15.0

# Optional
torch>=2.0.0
transformers>=4.30.0
